pyproject.toml
[project]
name = "nocaro"
version = "0.1.0"
description = ""
authors = [{ name = "Bobbyperson", email = "bobbyperson12345@gmail.com" }]
requires-python = ">=3.11,<3.13,!=3.12.5"
readme = "README.md"
license = "GNU GPLv3"
dependencies = [
    "aiosqlite>=0.21.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "pillow>=11.2.0",
    "discord.py[voice]>=2.5.2",
    "discord-pretty-help>=2.0.7",
    "requests>=2.32.3",
    "ffmpy>=0.5.0",
    "pyquery>=2.0.1",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "matplotlib>=3.10.1",
    "pydub>=0.25.1",
    "emoji>=2.14.1",
    "yfinance>=0.2.55",
    "pandas>=2.2.3",
    "pandas_market_calendars>=5.0.0",
    "pyttsx3>=2.98",
    "discord-ext-menus",
    "anyio>=4.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.4",
]

[tool.uv]
package = false

[tool.uv.sources]
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus" }

[tool.ruff.lint]
select = [
    "A004", # Check for import shadowing builtin functions: https://docs.astral.sh/ruff/rules/builtin-import-shadowing/
    "RUF",
    "UP",
    "I",
    "ASYNC",
    "F"
]
extend-ignore = [
    "ASYNC110", # I don't care
    "ASYNC221"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
